fix: release camera when QR scanner sheet closes
The three QR-scanner preview composables bound CameraX use cases to the
Activity lifecycle but never unbound them when the sheet was dismissed.
After any scan sheet was opened once, the Preview + ImageAnalysis use
cases stayed bound to the Activity lifecycle, so CameraX re-opened the
camera on every ON_START (app foreground, screen unlock), flashing the
Android camera privacy indicator even though no scan was requested.

Extract a shared QrCameraPreview composable that releases the camera
(unbindAll), shuts down the analysis executor, and closes the ML Kit
scanner on dispose, and replace the three duplicated previews with it.

Fixes the report in issue #65 comment about the camera indicator
flashing on app focus.
