+++ app/src/main/java/com/github/jvsena42/mandacaru/presentation/ui/components/QrCameraPreview.kt
@@ -1,0 +1,118 @@
+package com.github.jvsena42.mandacaru.presentation.ui.components
+
+import androidx.camera.core.CameraSelector
+import androidx.camera.core.ExperimentalGetImage
+import androidx.camera.core.ImageAnalysis
+import androidx.camera.core.ImageProxy
+import androidx.camera.core.Preview
+import androidx.camera.lifecycle.ProcessCameraProvider
+import androidx.camera.view.PreviewView
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberUpdatedState
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.viewinterop.AndroidView
+import androidx.core.content.ContextCompat
+import androidx.lifecycle.compose.LocalLifecycleOwner
+import com.google.mlkit.vision.barcode.BarcodeScanner
+import com.google.mlkit.vision.barcode.BarcodeScannerOptions
+import com.google.mlkit.vision.barcode.BarcodeScanning
+import com.google.mlkit.vision.barcode.common.Barcode
+import com.google.mlkit.vision.common.InputImage
+import java.util.concurrent.Executors
+import java.util.concurrent.atomic.AtomicReference
+
+@Composable
+fun QrCameraPreview(
+    enabled: Boolean,
+    onPayloadScanned: (String) -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    val lifecycleOwner = LocalLifecycleOwner.current
+    val lastPayload = remember { AtomicReference<String?>(null) }
+    val isEnabled = rememberUpdatedState(enabled)
+    val currentOnPayload = rememberUpdatedState(onPayloadScanned)
+    val cameraProvider = remember { AtomicReference<ProcessCameraProvider?>(null) }
+    val scanner = remember {
+        BarcodeScanning.getClient(
+            BarcodeScannerOptions.Builder()
+                .setBarcodeFormats(Barcode.FORMAT_QR_CODE)
+                .build()
+        )
+    }
+    val analysisExecutor = remember { Executors.newSingleThreadExecutor() }
+
+    LaunchedEffect(enabled) {
+        if (enabled) lastPayload.set(null)
+    }
+
+    DisposableEffect(Unit) {
+        onDispose {
+            cameraProvider.get()?.unbindAll()
+            analysisExecutor.shutdown()
+            scanner.close()
+        }
+    }
+
+    AndroidView(
+        modifier = modifier,
+        factory = { ctx ->
+            val previewView = PreviewView(ctx).apply {
+                scaleType = PreviewView.ScaleType.FILL_CENTER
+                implementationMode = PreviewView.ImplementationMode.COMPATIBLE
+            }
+            val providerFuture = ProcessCameraProvider.getInstance(ctx)
+            providerFuture.addListener({
+                val provider = providerFuture.get()
+                cameraProvider.set(provider)
+                val preview = Preview.Builder().build().also {
+                    it.setSurfaceProvider(previewView.surfaceProvider)
+                }
+                val analysis = ImageAnalysis.Builder()
+                    .setBackpressureStrategy(ImageAnalysis.STRATEGY_KEEP_ONLY_LATEST)
+                    .build()
+                    .also {
+                        it.setAnalyzer(analysisExecutor) { proxy ->
+                            processFrame(proxy, scanner) { payload ->
+                                if (isEnabled.value && lastPayload.getAndSet(payload) != payload) {
+                                    currentOnPayload.value(payload)
+                                }
+                            }
+                        }
+                    }
+                try {
+                    provider.unbindAll()
+                    provider.bindToLifecycle(
+                        lifecycleOwner,
+                        CameraSelector.DEFAULT_BACK_CAMERA,
+                        preview,
+                        analysis,
+                    )
+                } catch (_: IllegalStateException) {
+                }
+            }, ContextCompat.getMainExecutor(ctx))
+            previewView
+        },
+    )
+}
+
+@androidx.annotation.OptIn(ExperimentalGetImage::class)
+private fun processFrame(
+    proxy: ImageProxy,
+    scanner: BarcodeScanner,
+    onPayload: (String) -> Unit,
+) {
+    val mediaImage = proxy.image
+    if (mediaImage == null) {
+        proxy.close()
+        return
+    }
+    val image = InputImage.fromMediaImage(mediaImage, proxy.imageInfo.rotationDegrees)
+    scanner.process(image)
+        .addOnSuccessListener { barcodes ->
+            barcodes.firstOrNull { it.rawValue != null }?.rawValue?.let(onPayload)
+        }
+        .addOnCompleteListener { proxy.close() }
+}
